requirements.txt
asgiref==3.3.4
Django==3.2
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-phonenumber-field==5.1.0
mysqlclient==2.0.3
phonenumbers==8.12.21
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1